// src/modules/chat/enums/chat.enum.ts
export enum ChatTypeEnum {
  PRIVATE = 'PRIVATE',
  GROUP = 'GROUP',
}

export enum SubScriptionTopicsEnum {
  ONLINE_STATUS = 'ONLINE_STATUS',
  TYPING_STATUS = 'TYPING_STATUS',
  CREATE_MESSAGE = 'CREATE_MESSAGE',
  UPDATE_MESSAGE = 'UPDATE_MESSAGE',
  DELETE_MESSAGE = 'DELETE_MESSAGE',
  DELIVERED_MESSAGE = 'DELIVERED_MESSAGE',
  SEEN_MESSAGE = 'SEEN_MESSAGE',
}

export enum ClientEventsEnum {
  MY_TYPING_STATUS = 'MY_TYPING_STATUS',
  MY_SEEN_MESSAGE = 'MY_SEEN_MESSAGE',
}